requirements.txt
flask>=3.0
pyjwt>=2.8
bcrypt>=4.1
pyotp>=2.9
qrcode>=7.4
python-dotenv>=1.0
flask-cors>=4.0
bleach>=6.1
pillow>=10.0
gunicorn>=21.2
openpyxl>=3.1
